Updated meshing configurations
Updated the bed meshing and probe sections with more aggressive tolerance requirements and mode probe points. It will take longer, but the results is a more accurate mesh, which is very helpful for a bed this size and an inductive prove. When using KAMP it will only probe the printable area anyway, so it wont take long for small objects